refactor: create a single file for single tone constants
